Migrate index config to another file to clean up the main helper function
