+++ indexer/index-manager/src/helper.rs
@@ -23,12 +23,11 @@ use index_store::core::IndexStore;
 use ipfs_client::core::create_ipfs_clients;
 use plugin::manager::PluginManager;
 use stream_mod::{HelloRequest, GetBlocksRequest, GenericDataProto, ChainType, DataType, streamout_client::StreamoutClient};
-
+use crate::builder::{IndexConfigLocalBuilder, IndexConfigIpfsBuilder};
 // Refactor to new files for substrate / solana
 use massbit_chain_substrate::data_type::{decode, decode_transactions, SubstrateBlock as Block, SubstrateBlock, SubstrateHeader as Header, SubstrateUncheckedExtrinsic as Extrinsic, get_extrinsics_from_block, SubstrateEventRecord};
 use massbit_chain_solana::data_type::{SolanaBlock, decode as solana_decode, SolanaEncodedBlock, convert_solana_encoded_block_to_solana_block, SolanaTransaction, SolanaLogMessages};
-use std::sync::Arc;
-use crate::builder::{IndexConfigLocalBuilder, IndexConfigIpfsBuilder};
+
 
 // Configs
 pub mod stream_mod {
@@ -58,87 +57,6 @@ pub async fn get_index_config(ipfs_config_hash: &String) -> serde_yaml::Mapping 
         .to_vec();
 
     serde_yaml::from_slice(&file_bytes).unwrap()
-}
-
-pub async fn get_mapping_file_from_ipfs(ipfs_mapping_hash: &String) -> String {
-    let ipfs_addresses = vec![IPFS_ADDRESS.to_string()];
-    let ipfs_clients = create_ipfs_clients(&ipfs_addresses).await; // Refactor to use lazy load
-
-    let file_bytes = ipfs_clients[0]
-        .cat_all(ipfs_mapping_hash.to_string())
-        .compat()
-        .await
-        .unwrap()
-        .to_vec();
-
-    let file_name = [ipfs_mapping_hash, ".so"].join("");
-    let res = fs::write(&file_name, file_bytes); // Add logger and says that write file successfully
-
-    match res {
-        Ok(_) => {
-            log::info!("[Index Manager Helper] Write SO file to local storage successfully");
-            file_name
-        }
-        Err(err) => {
-            panic!(
-                "[Index Manager Helper] Could not write file to local storage {:#?}",
-                err
-            )
-        }
-    }
-}
-
-pub async fn get_config_file_from_ipfs(ipfs_config_hash: &String) -> String {
-    let ipfs_addresses = vec![IPFS_ADDRESS.to_string()];
-    let ipfs_clients = create_ipfs_clients(&ipfs_addresses).await; // Refactor to use lazy load
-
-    let file_bytes = ipfs_clients[0]
-        .cat_all(ipfs_config_hash.to_string())
-        .compat()
-        .await
-        .unwrap()
-        .to_vec();
-
-    let file_name = [ipfs_config_hash, ".yaml"].join("");
-    let res = fs::write(&file_name, file_bytes); // Add logger and says that write file successfully
-
-    match res {
-        Ok(_) => {
-            log::info!(
-                "[Index Manager Helper] Write project.yaml file to local storage successfully"
-            );
-            file_name
-        }
-        Err(err) => {
-            panic!(
-                "[Index Manager Helper] Could not write file to local storage {:#?}",
-                err
-            )
-        }
-    }
-}
-
-
-
-pub fn get_mapping_file_from_local(mapping_path: &String) -> PathBuf {
-    let so_file_path = PathBuf::from(mapping_path.to_string());
-    so_file_path
-}
-
-pub fn get_config_file_from_local(config_path: &String) -> String {
-    let mut config_file = String::new();
-    let mut f = File::open(config_path).expect("Unable to open file");
-    f.read_to_string(&mut config_file)
-        .expect("Unable to read string");
-    config_file
-}
-
-pub fn get_raw_query_from_local(model_path: &String) -> String {
-    let mut raw_query = String::new();
-    let mut f = File::open(model_path).expect("Unable to open file");
-    f.read_to_string(&mut raw_query)
-        .expect("Unable to read string");
-    raw_query
 }
 
 pub fn create_new_indexer_detail_table(connection: &PgConnection, raw_query: &String) {
@@ -214,7 +132,7 @@ pub async fn track_hasura_table(table_name: &String) {
 pub async fn loop_blocks(params: DeployParams) -> Result<(), Box<dyn Error>> {
     let mut store = IndexStore::new(DATABASE_CONNECTION_STRING.as_str());
 
-    // Get mapping file, raw query to create new table and project.yaml config
+    // Get user index mapping logic, query for migration and index's configurations
     let index_config = match params.deploy_type {
         DeployType::Local => {
             let index_config = IndexConfigLocalBuilder::default()
